[#123] storing flow-config in atom and files in subdir
As the flow-server-config can be updated at runtime, it is not longer
immutable, so we need to store it in an atom.

Also, we store the p12 and appengine.xml file in disk, so when there is
a refresh, they need to be refetched.

We opted to store the files in a subdirectory with the schema $tmpdir/$refresh-timestamp,
so that if there is another thread using the old ones, it cannot be impacted by the refresh.