[#134] Use the combination of hostname + timestamp for a heap dump
We can configure the JVM to generate a heap dump when an
OutOfMemoryError happens.  The parameter -XX:HeapDumpPath can be a
folder or a file.

When is a folder the JVM generates a java_pid<pid>.hprof file.
In the case of our services the java process running in a container we
always have the PID 1, which means that the path to the file is the
same.

The JVM will refuse to overwrite a file if this already exists,
e.g. Unable to create /dumps/java_pid1.hprof: File exists

By concatenating the container hostname and timestamp we generate
a unique filename for the heap dump.