[#148] Auth0 specific container
Auth0 does not provide an introspect endpoint so we validate the JWT
access token and then call the /userinfo endpoint to get the email
address.

Another option could be to expect an id_token, but as per
https://auth0.com/docs/api-auth/why-use-access-tokens-to-secure-apis
is not the most secure way.

The user info is cached for 15 minutes.